# docs/swagger.yaml
basePath: /
definitions:
  erg_ninja_internal_dto_response.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      path:
        type: string
      request_id:
        type: string
      statusCode:
        type: integer
      timestamp:
        type: string
    type: object
...
  erg_ninja_internal_modules_audit_dto.AuditLogResponse:
    properties:
      action:
        type: string
      changes:
        type: string
      createdAt:
        type: string
      id:
        type: string
      ip_address:
        type: string
      metadata:
        type: string
      resource_id:
        type: string
      resource_type:
        type: string
      user_agent:
        type: string
      user_email:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: ERG Support
    url: https://erg.ninja/support
  description: The unified backend API for ERG — a multi-tenant eLearning, content
    management, and automation platform.
  title: ERG Platform API
  version: "1.0"
paths:
...
  /api/audit/logs:
    get:
      description: Returns a paginated list of audit logs with filters.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by action
        in: query
        name: action
        type: string
      - description: Filter by user
        in: query
        name: user_id
        type: string
      - description: Filter by resource
        in: query
        name: resource_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/erg_ninja_internal_dto_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/erg_ninja_internal_modules_audit_dto.AuditLogResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List audit logs
      tags:
      - Audit
  /api/audit/logs/{id}:
    get:
      description: Returns the details of a single audit log entry.
      parameters:
      - description: Log ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/erg_ninja_internal_modules_audit_dto.AuditLogResponse'
      security:
      - BearerAuth: []
      summary: Get audit log detail
      tags:
      - Audit
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer <token>" to authenticate.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"